Refactor to avoid dynamic handler loading, and single data-class files.
