--- zyngine/ctrldev/akai_apc_key25/brights.py
@@ -1,28 +1,0 @@
-class BRIGHTS: 
-    LED_OFF = 0x80
-    LED_BRIGHT_10 = 0x90
-    LED_BRIGHT_25 = 0x91
-    LED_BRIGHT_50 = 0x92
-    LED_BRIGHT_65 = 0x93
-    LED_BRIGHT_75 = 0x94
-    LED_BRIGHT_90 = 0x95
-    LED_BRIGHT_100 = 0x96
-    LED_PULSING_16 = 0x97
-    LED_PULSING_8 = 0x98
-    LED_PULSING_4 = 0x99
-    LED_PULSING_2 = 0x9A
-    LED_BLINKING_24 = 0x9B
-    LED_BLINKING_16 = 0x9C
-    LED_BLINKING_8 = 0x9D
-    LED_BLINKING_4 = 0x9E
-    LED_BLINKING_2 = 0x9F
-
-LED_BRIGHTS = [
-    BRIGHTS.LED_BRIGHT_10,
-    BRIGHTS.LED_BRIGHT_25,
-    BRIGHTS.LED_BRIGHT_50,
-    BRIGHTS.LED_BRIGHT_65,
-    BRIGHTS.LED_BRIGHT_75,
-    BRIGHTS.LED_BRIGHT_90,
-    BRIGHTS.LED_BRIGHT_100,
-]
