--- zyngine/ctrldev/akai_apc_key25/buttons.py
@@ -1,45 +1,0 @@
-# APC Key25 buttons
-class BUTTONS:
-    BTN_SHIFT = 0x62
-    BTN_STOP_ALL_CLIPS = 0x51
-    BTN_PLAY = 0x5B
-    BTN_RECORD = 0x5D
-
-    BTN_TRACK_1 = BTN_UP = 0x40
-    BTN_TRACK_2 = BTN_DOWN = 0x41
-    BTN_TRACK_3 = BTN_LEFT = BTN_UNDO = 0x42
-    BTN_TRACK_4 = BTN_RIGHT = BTN_REDO = 0x43
-    BTN_TRACK_5 = BTN_KNOB_CTRL_VOLUME = 0x44
-    BTN_TRACK_6 = BTN_KNOB_CTRL_PAN = 0x45
-    BTN_TRACK_7 = BTN_KNOB_CTRL_SEND = 0x46
-    BTN_TRACK_8 = BTN_KNOB_CTRL_DEVICE = 0x47
-
-    BTN_SOFT_KEY_START = 0x52
-    BTN_SOFT_KEY_END = 0x56
-    BTN_SOFT_KEY_CLIP_STOP = BTN_KNOB_1 = 0x52
-    BTN_SOFT_KEY_SOLO = BTN_KNOB_2 = 0x53
-    BTN_SOFT_KEY_MUTE = BTN_KNOB_3 = 0x54
-    BTN_SOFT_KEY_REC_ARM = BTN_KNOB_4 = 0x55
-    BTN_SOFT_KEY_SELECT = 0x56
-
-    BTN_PAD_START = 0x00
-    BTN_PAD_END = 0x27
-    BTN_PAD_29 = BTN_ALT = 0x1C
-    BTN_PAD_30 = BTN_METRONOME = 0x1D
-    BTN_PAD_31 = BTN_PAD_STEP = 0x1E
-    BTN_PAD_37 = BTN_OPT_ADMIN = 0x24
-    BTN_PAD_38 = BTN_MIX_LEVEL = 0x25
-    BTN_PAD_39 = BTN_CTRL_PRESET = 0x26
-    BTN_PAD_40 = BTN_ZS3_SHOT = 0x27
-    BTN_PAD_5 = BTN_PAD_LEFT = 0x04
-    BTN_PAD_6 = BTN_PAD_DOWN = 0x05
-    BTN_PAD_7 = BTN_PAD_RIGHT = 0x06
-    BTN_PAD_8 = BTN_F4 = 0x07
-    BTN_PAD_13 = BTN_BACK_NO = 0x0C
-    BTN_PAD_14 = BTN_PAD_UP = 0x0D
-    BTN_PAD_15 = BTN_SEL_YES = 0x0E
-    BTN_PAD_16 = BTN_F3 = 0x0F
-    BTN_PAD_21 = BTN_PAD_RECORD = 0x14
-    BTN_PAD_23 = BTN_PAD_PLAY = 0x16
-    BTN_PAD_24 = BTN_F2 = 0x17
-    BTN_PAD_32 = BTN_F1 = 0x1F
