--- zyngine/ctrldev/akai_apc_key25/colors.py
@@ -1,31 +1,0 @@
-class COLORS:
-    COLOR_BLACK = 0x00
-    COLOR_DARK_GREY = 0x01
-    COLOR_RED = 0x05
-    COLOR_GREEN = 0x15
-    COLOR_BLUE = 0x25
-    COLOR_AQUA = 0x21
-    COLOR_BLUE_DARK = 0x2D
-    COLOR_BLUE_LIGHT = 0x24
-    COLOR_WHITE = 0x03
-    COLOR_EGYPT = 0x6C
-    COLOR_ORANGE = 0x09
-    COLOR_ORANGE_LIGHT = 0x08
-    COLOR_AMBER = 0x54
-    COLOR_RUSSET = 0x3D
-    COLOR_PURPLE = 0x51
-    COLOR_PINK = 0x39
-    COLOR_PINK_LIGHT = 0x52
-    COLOR_PINK_WARM = 0x38
-    COLOR_YELLOW = 0x0D
-    COLOR_LIME = 0x4B
-    COLOR_LIME_DARK = 0x11
-    COLOR_DARK_GREEN = 0x41
-    COLOR_GREEN_YELLOW = 0x4A
-    COLOR_BROWNISH_RED = 0x0A
-    COLOR_BROWN_LIGHT = 0x7E
-    SOFT_OFF = 0x00
-    SOFT_ON = 0x01
-    SOFT_BLINK = 0x02
-
-
